deleted extra content, added favicon, added env variables for increased security
