+++ culinary-compass/app/layout.js
@@ -4,8 +4,14 @@ import "./globals.css";
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "Culinary Compass",
+  description: "Your ultimate travel companion! Plan your culinary adventures with ease as CulinaryCompass maps out the perfect path between restaurants, ensuring you savor every moment of your vacation.",
+  icons:{
+    icon:['/favicon.ico?v=4'],
+    apple:['/apple-touch-icon.png?v=4'],
+    shortcut:['apple-touch-icon.png'],
+  },
+  manifest: '/site.webmanifest'
 };
 
 export default function RootLayout({ children }) {
