+++ culinary-compass/next.config.mjs
@@ -1,4 +1,8 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+    env: {
+        BASE_URL: process.env.BASE_URL,
+      }
+};
 
 export default nextConfig;
