+++ culinary-compass/package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@react-google-maps/api": "^2.19.3",
     "d3": "^7.9.0",
+    "dotenv": "^16.4.5",
     "graph-data-structure": "^3.5.0",
     "leaflet": "^1.9.4",
     "next": "^14.2.2",
